requirements.txt
# GEO Agent 项目依赖

# Python 版本要求
# Python >= 3.10

# Agent 框架
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-vertexai>=0.0.6

# 异步 HTTP 客户端
aiohttp>=3.9.0
httpx>=0.25.0

# 数据库
pymongo>=4.6.0
motor>=3.3.0  # 异步 MongoDB 驱动

# API 客户端
openai>=1.12.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
tweepy>=4.14.0  # X API (Grok)

# NER 工具
spacy>=3.7.0

# 环境变量管理
python-dotenv>=1.0.0

# 数据处理
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 日志和工具
structlog>=24.1.0

# API 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

